+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Notable changes will be documented in this file. The project adheres to [Semanti
 Unreleased
 ----------
 
+* Improved `Event` Grid Element
 * Updated TypoScript (Fluid Styled Content)
 
 4.2.0