+++ CHANGELOG.md
@@ -3,6 +3,12 @@ Changelog
 
 Notable changes will be documented in this file. The project adheres to [Semantic Versioning].
 
+Unreleased
+-----------
+
+* Updated dependencies
+* Updated constraints to TYPO3 7.6
+
 4.4.0
 -----